Create 55. Jump Game
Initially tried to reach the last element of the array starting from the first. It does not address all the edge cases until you implement decision trees or cache which are time consuming. 
The provided solution starting from the last element and shifts the goal to the previous elements if they can reach the goal which reduces the time complexity to O(n)