Create 383. Ransom Note
Here we use hash maps to solve the given problem.
Time complexity: O(m+n)
Space complexity: O(1)
Here, we first create an empty dict and we add the frequency of each character of magazine in the hash map
we then go through each character from the ransom list and if the character matches to the one in our dict, we decrease the frequency of that char by 1. If we fail to do so, we return false, or else true.