Create 205. Isomorphic Strings
This solution creates to empty hash maps and we compare each character of one string to the other (and the second string to the first). If there is a same character, we see if it is mapped to some other character or not. If it is, we return False, else True