Update 290. Word Pattern
The updated solution maps the chars to words and the words to chars. It further checks if the chars and words match each other or not and returns False if there exists another char/word already mapped to a word/char. It also addresses some edge cases where it returns False if the length of both the lists is not same