Create 1. Two Sum
This solution first creates a hash map consisting of all the arr values as keys and all the solution to those keys as values. we check if we have the key solutions in the arr and return their indexes. But this solution does not work if there are duplicate elements in an array