Update 1. Two Sum
This solution creates a hash map of all the elements that are iterated through the list along with their indices. When we find our second element for the solution, we look up to the value of the first element from our hash map and return both the indices. The time complexity of this solution is O(n) and the space complexity is O(n) as well.