Create 202. Happy Number
Memory complexity: O(n)
This problem can also be solved with the help of linked lists. This solution basically saves the sum of the digits in a set if it is new and iterates through the loop endlessly until you either get a duplicate sum (resulting in looping within the small loop again = false) or looping within itself with one as the sum (1 squared is 1 itself = True). We return True if the sum reaches one or else we return false. We also construct another function named sumOfSquares to calculate the sum of squares of all the digits of the number. We use the mod and division functions to achieve this.