Create 12. Integer to Roman
We first create a dictionary (in descending order) from the largest value to the smallest (order is important as we will iterate through the hmap from the first value pair. We also create the value symbol pairs for the special numbers from 4 and 9 (like 40, 90, 400, 900) in our hash map

now, we iterate through the key value pairs in our hash map while our number is greater than the greatest value we are iterating through. we further add the symbol of the greatest value and subtract it from our result.